git log --pretty=format:"%h %ad | %s%d [%an]" origin/abt/topic/gmin/l-dev/mr1/master..origin/abt/topic/gmin/l-dev/mr1/audio/master:
  6f3b259 Wed Mar 11 15:31:45 2015 +0100 | [AUDIO] 64 bit issue: use size_t instead of uint32_t in audio_stream_manager (origin/abt/topic/gmin/l-dev/mr1/audio/master) [Frederic Boisnard]
  7b9b960 Wed Mar 11 13:47:04 2015 +0100 | Revert "Audio: Fix 64bit compilation issue of audio_stream_manager" [Frederic Boisnard]
  f2d8ec1 Tue Dec 9 17:16:10 2014 +0100 | Allow RogueElement to array conversion. [Piotr Diop]
  d97a354 Tue Dec 9 16:56:17 2014 +0100 | Allow float type usage for rogue parameters [Piotr Diop]
  f12d799 Tue Nov 18 16:18:51 2014 +0100 | Manage the count of written frames within AudioHAL [Francois Gaffie]
  4a210c0 Fri Sep 12 13:53:31 2014 +0200 | Remove LOCAL_IMPORT_C_INCLUDE_DIRS_FROM_XXX_LIBRARIES [pafonsoX]
  4883ec8 Wed Dec 17 09:46:12 2014 +0100 | Fix output audio device used in Audio HAL hardware device test [Apelete Seketeli]
  f72b698 Thu Oct 23 14:20:36 2014 +0200 | 64-bit: fix types to build for 64-bit targets [Francois Gaffie]
  ed96fa3 Fri Oct 3 17:36:52 2014 +0200 | Update ModemProxy with latest MAMGR IF APIs [Francois Gaffie]
  c2701a8 Fri Oct 3 11:16:06 2014 +0200 | Update Modem Proxy to get more precise call status [Francois Gaffie]

git diff --stat origin/abt/topic/gmin/l-dev/mr1/master..HEAD:
  audio_conversion/src/AudioReformatter.cpp          |  4 +-
  audio_conversion/src/AudioRemapper.cpp             |  2 +-
  audio_conversion/test/AudioConversionTest.cpp      | 10 ++--
  audio_platform_state/config/route_criteria.conf    |  5 +-
  audio_platform_state/src/AudioHalConf.hpp          |  1 +
  audio_platform_state/src/AudioPlatformState.cpp    |  7 +++
  audio_platform_state/src/ModemProxy.cpp            | 19 ++++---
  audio_platform_state/src/ModemProxy.hpp            | 20 +++----
  audio_route_manager/Android.mk                     |  7 +--
  audio_route_manager/AudioRouteManager.hpp          | 10 ++++
  audio_route_manager/interface/IStreamInterface.hpp | 36 ++++++++++---
  audio_stream_manager/Android.mk                    |  7 +--
  audio_stream_manager/src/StreamIn.cpp              |  2 +-
  audio_stream_manager/src/StreamOut.cpp             |  8 ++-
  hardware_device/test/DeviceTest.cpp                |  8 +--
  parameter_mgr_helper/ParameterMgrHelper.cpp        | 61 +++++++++++++++++++++-
  stream_lib/Android.mk                              |  7 ++-
  stream_lib/TinyAlsaIoStream.cpp                    | 14 +++--
  stream_lib/include/IoStream.hpp                    |  2 +-
  stream_lib/include/TinyAlsaIoStream.hpp            |  2 +-
  20 files changed, 170 insertions(+), 62 deletions(-)

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4991

+++ audio_conversion/src/AudioReformatter.cpp
@@ -70,7 +70,7 @@ status_t AudioReformatter::convertS16toS24over32(const void *src,
                                                  const size_t inFrames,
                                                  size_t *outFrames)
 {
-    uint32_t frameIndex;
+    size_t frameIndex;
     const int16_t *src16 = (const int16_t *)src;
     uint32_t *dst32 = (uint32_t *)dst;
     size_t n = inFrames * mSsSrc.getChannelCount();
@@ -94,7 +94,7 @@ status_t AudioReformatter::convertS24over32toS16(const void *src,
 {
     const uint32_t *src32 = (const uint32_t *)src;
     int16_t *dst16 = (int16_t *)dst;
-    uint32_t frameIndex;
+    size_t frameIndex;
 
     size_t n = inFrames * mSsSrc.getChannelCount();
