+++ audio_stream_manager/src/StreamIn.cpp
@@ -636,7 +636,7 @@ bool StreamIn::isAecEffect(effect_handle_t effect)
 void StreamIn::getCaptureDelay(struct echo_reference_buffer *buffer)
 {
     /* read frames available in kernel driver buffer */
-    uint32_t kernel_frames;
+    size_t kernel_frames;
     struct timespec tstamp;
     long buf_delay;
     long kernel_delay;
